requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
gevent==23.9.1
gevent-websocket==0.10.1
Werkzeug==3.0.1
qrcode==7.4.2
Pillow==11.2.1
click==8.1.7
pytest==7.4.3
setuptools>=68.0.0